packages/nvidia_nat_phoenix/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64", "setuptools-scm>=8"]


[tool.setuptools.packages.find]
where = ["src"]
include = ["nat.*"]


[tool.setuptools_scm]
git_describe_command = "git describe --long --first-parent"
root = "../.."


[project]
name = "nvidia-nat-phoenix"
dynamic = ["version"]
dependencies = [
  # Keep package version constraints as open as possible to avoid conflicts with other packages. Always define a minimum
  # version when adding a new package. If unsure, default to using `~=` instead of `==`. Does not apply to nvidia-nat packages.
  # Keep sorted!!!
  "nvidia-nat[opentelemetry]~=1.3",
  "arize-phoenix-otel~=0.13.1",
  "openinference-instrumentation", # version determined by arize-phoenix-otel
]
requires-python = ">=3.11,<3.14"
description = "Subpackage for Arize Phoenix integration in NeMo Agent toolkit"
readme = "src/nat/meta/pypi.md"
keywords = ["ai", "observability", "phoenix", "arize"]
license = { text = "Apache-2.0" }
authors = [{ name = "NVIDIA Corporation" }]
maintainers = [{ name = "NVIDIA Corporation" }]
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]

[project.urls]
documentation = "https://docs.nvidia.com/nemo/agent-toolkit/latest/"
source = "https://github.com/NVIDIA/NeMo-Agent-Toolkit"


[tool.uv]
config-settings = { editable_mode = "compat" }


[tool.uv.sources]
nvidia-nat = { workspace = true }


[project.entry-points.'nat.components']
nat_phoenix = "nat.plugins.phoenix.register"
